Move frets into their own object
Will interact better with the in-progress pedal/lever object
