+++ lib/copedent.rb
@@ -3,6 +3,7 @@
 require_relative "copedent/version"
 require_relative "copedent/sheet"
 require_relative "copedent/printer"
+require_relative "copedent/fret"
 require "terminal-table"
 require "csv"
 
@@ -12,26 +13,16 @@ require "csv"
 # TODO add a hash or first-class object for frets
 # TODO need to choose a CLI tools library
 
-# holds everything for now, will break out into more classes soon
 module Copedent
-  # === STATIC CONFIG ===
+  class CopedentError < StandardError; end
 
   ALL_TYPES = [:major, :minor, :dom7] # unused, for reference
 
   ALL_NOTES = ["E", "F", "F#", "G", "G#", "A", "A#", "B", "C", "C#", "D", "D#"]
 
-  # TODO - look for multiple combinations of the triad
-  QUALITIES = {
-    major: ["3"],
-    minor: ["b3"],
-    dom7: ["3", "b7"]
-  }
-
-  # TODO: use the 9+ qualities for QoL
-  RELATIONS = [
-    "Root", "b2", "2", "b3", "3",
-    "4", "b5", "5", "#5", "6", "b7",
-    "maj7", "Root", "b9", "9", "#9",
-    "11", "#11", "b13", "13"
-  ]
+  # TODO: allow passing in aliases (e.g. :a for :p1)
+  # TODO: these should all be passed in?
+  PEDAL_NAMES = [:p0, :p1, :p2, :p3, :p4, :p5, :p6, :p7, :p8]
+  LEVER_NAMES = [:lkl1, :lkl2, :lkv, :lkr, :rkl, :rkr]
+  ALL_MODS = PEDAL_NAMES | LEVER_NAMES
 end
